refactor: separate unit tests from integration tests in pkg/paths
- Move git_discovery_test.go entirely to integration_test.go (all git operations)
- Split validation_test.go:
  - Keep pure logic tests (path validation, concurrent access, pack name validation)
  - Move filesystem and environment tests to validation_integration_test.go
- Integration tests moved:
  - TestPathSecurityValidation (creates temp directories)
  - TestEnvironmentVariableHandling (manipulates OS environment)
  - All git discovery tests (git commands, directory changes)

This maintains the pattern where:
- Unit tests verify pure logic without external dependencies
- Integration tests handle filesystem, environment, and git operations
