+++ pkg/paths/validation_test.go
@@ -11,6 +11,8 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+// TestPathValidation tests path validation logic
+// This is a unit test - it tests pure validation logic without filesystem operations
 func TestPathValidation(t *testing.T) {
 	p, err := New("/test/dotfiles")
 	testutil.AssertNoError(t, err)
@@ -70,153 +72,8 @@ func TestPathValidation(t *testing.T) {
 	}
 }
 
-func TestPathSecurityValidation(t *testing.T) {
-	// Create a temporary directory structure for testing
-	tmpRoot := testutil.TempDir(t, "path-security")
-	dotfilesDir := filepath.Join(tmpRoot, "dotfiles")
-	testutil.CreateDir(t, tmpRoot, "dotfiles")
-	testutil.CreateDir(t, tmpRoot, "outside")
-
-	p, err := New(dotfilesDir)
-	testutil.AssertNoError(t, err)
-
-	tests := []struct {
-		name           string
-		path           string
-		shouldBeInside bool
-		description    string
-	}{
-		{
-			name:           "file inside dotfiles",
-			path:           filepath.Join(dotfilesDir, "vim", "vimrc"),
-			shouldBeInside: true,
-			description:    "Files inside dotfiles should be detected",
-		},
-		{
-			name:           "dotfiles root itself",
-			path:           dotfilesDir,
-			shouldBeInside: true,
-			description:    "Dotfiles root should be considered inside",
-		},
-		{
-			name:           "file outside dotfiles",
-			path:           filepath.Join(tmpRoot, "outside", "file.txt"),
-			shouldBeInside: false,
-			description:    "Files outside dotfiles should be detected",
-		},
-		{
-			name:           "path traversal attempt",
-			path:           filepath.Join(dotfilesDir, "..", "outside", "file.txt"),
-			shouldBeInside: false,
-			description:    "Path traversal should be detected",
-		},
-		{
-			name:           "symlink target outside (path only)",
-			path:           "/etc/passwd",
-			shouldBeInside: false,
-			description:    "System files should be outside dotfiles",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			isInside, err := p.IsInDotfiles(tt.path)
-			testutil.AssertNoError(t, err)
-			testutil.AssertEqual(t, tt.shouldBeInside, isInside)
-		})
-	}
-}
-
-func TestEnvironmentVariableHandling(t *testing.T) {
-	// Save original environment
-	origDotfilesRoot := os.Getenv(EnvDotfilesRoot)
-	origDodotDataDir := os.Getenv(EnvDodotDataDir)
-	origDodotConfigDir := os.Getenv(EnvDodotConfigDir)
-	origDodotCacheDir := os.Getenv(EnvDodotCacheDir)
-	origXdgStateHome := os.Getenv("XDG_STATE_HOME")
-
-	t.Cleanup(func() {
-		_ = os.Setenv(EnvDotfilesRoot, origDotfilesRoot)
-		_ = os.Setenv(EnvDodotDataDir, origDodotDataDir)
-		_ = os.Setenv(EnvDodotConfigDir, origDodotConfigDir)
-		_ = os.Setenv(EnvDodotCacheDir, origDodotCacheDir)
-		_ = os.Setenv("XDG_STATE_HOME", origXdgStateHome)
-	})
-
-	tests := []struct {
-		name     string
-		envSetup map[string]string
-		validate func(t *testing.T, p *Paths)
-	}{
-		{
-			name: "DOTFILES_ROOT with spaces",
-			envSetup: map[string]string{
-				EnvDotfilesRoot: "/path with spaces/dotfiles",
-			},
-			validate: func(t *testing.T, p *Paths) {
-				testutil.AssertEqual(t, "/path with spaces/dotfiles", p.DotfilesRoot())
-			},
-		},
-		{
-			name: "DOTFILES_ROOT with tilde",
-			envSetup: map[string]string{
-				EnvDotfilesRoot: "~/my-dotfiles",
-			},
-			validate: func(t *testing.T, p *Paths) {
-				homeDir, _ := os.UserHomeDir()
-				expected := filepath.Join(homeDir, "my-dotfiles")
-				testutil.AssertEqual(t, expected, p.DotfilesRoot())
-			},
-		},
-		{
-			name: "Custom XDG_STATE_HOME",
-			envSetup: map[string]string{
-				"XDG_STATE_HOME": "/custom/state",
-			},
-			validate: func(t *testing.T, p *Paths) {
-				expected := filepath.Join("/custom/state", "dodot", "dodot.log")
-				testutil.AssertEqual(t, expected, p.LogFilePath())
-			},
-		},
-		{
-			name: "All custom directories",
-			envSetup: map[string]string{
-				EnvDodotDataDir:   "/custom/data/dodot",
-				EnvDodotConfigDir: "/custom/config/dodot",
-				EnvDodotCacheDir:  "/custom/cache/dodot",
-			},
-			validate: func(t *testing.T, p *Paths) {
-				testutil.AssertEqual(t, "/custom/data/dodot", p.DataDir())
-				testutil.AssertEqual(t, "/custom/config/dodot", p.ConfigDir())
-				testutil.AssertEqual(t, "/custom/cache/dodot", p.CacheDir())
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Clear all relevant env vars
-			_ = os.Unsetenv(EnvDotfilesRoot)
-			_ = os.Unsetenv(EnvDodotDataDir)
-			_ = os.Unsetenv(EnvDodotConfigDir)
-			_ = os.Unsetenv(EnvDodotCacheDir)
-			_ = os.Unsetenv("XDG_STATE_HOME")
-
-			// Set up test environment
-			for k, v := range tt.envSetup {
-				_ = os.Setenv(k, v)
-			}
-
-			p, err := New("")
-			testutil.AssertNoError(t, err)
-
-			if tt.validate != nil {
-				tt.validate(t, p)
-			}
-		})
-	}
-}
-
+// TestCrossPlatformPaths tests cross-platform path handling
+// This is a unit test - it tests pure path manipulation logic
 func TestCrossPlatformPaths(t *testing.T) {
 	p, err := New("/test/dotfiles")
 	testutil.AssertNoError(t, err)
@@ -264,6 +121,8 @@ func TestCrossPlatformPaths(t *testing.T) {
 	}
 }
 
+// TestPathExpansionEdgeCases tests tilde expansion edge cases
+// This is a unit test with minimal OS interaction (just for home directory)
 func TestPathExpansionEdgeCases(t *testing.T) {
 	tests := []struct {
 		name     string
@@ -328,6 +187,8 @@ func TestPathExpansionEdgeCases(t *testing.T) {
 	}
 }
 
+// TestDeploymentPathStructure tests deployment path structure
+// This is a unit test - it tests pure path computation logic
 func TestDeploymentPathStructure(t *testing.T) {
 	p, err := New("/test/dotfiles")
 	testutil.AssertNoError(t, err)
@@ -367,6 +228,8 @@ func TestDeploymentPathStructure(t *testing.T) {
 	})
 }
 
+// TestPathsConcurrentAccess tests concurrent access to paths
+// This is a unit test - it tests thread safety without filesystem operations
 func TestPathsConcurrentAccess(t *testing.T) {
 	// Test that a single Paths instance can be safely accessed concurrently
 	// This is what actually matters for our use case
@@ -418,6 +281,8 @@ func TestPathsConcurrentAccess(t *testing.T) {
 	}
 }
 
+// TestValidatePackName tests pack name validation
+// This is a pure unit test - no filesystem operations
 func TestValidatePackName(t *testing.T) {
 	tests := []struct {
 		name    string
